Hmm... an if-template is needed. In members there can be any declaration
"A non-static data member has semantics different from that of ordinary variable and static data-members. Consequently, it is represented by a distinct node (Field). However, the steps
for manifacturing such node is very similar to that of a variable, except that the construction
function declare_field() is used -- instead of declare_var()."